esm2_model_embeddings/requirements.txt
torch
transformers
biopython
numpy